hygiene(tests): relocate 20 CLI/banner test files to tests/unit/dx/
Files that import from tolokaforge.dx.* belong under tests/unit/dx/
so a headless install (`pip install tolokaforge`, no [dx] extras) can
collect and run the engine test tier without pulling the dx surface.
Reinforces ADR-0019's front-end plugin boundary at the test tree.

Moved:
- 6 files originally listed in #456: test_cli_stdout_contract,
  test_cli_model_cost_config, test_cli_budget_flags,
  test_logging_cli_flags, test_cli_status, test_run_display_cost_meter.
- 14 additional CLI/banner test files surfaced by a broader audit:
  test_cli_display_flag, test_cli_display, test_cli_grouped_help,
  test_cli_assets, test_cli_resume, test_cli_commands,
  test_cli_fallback_flag, test_cli_budget_integration,
  test_dry_run_cli, test_adapter_convert_cli,
  test_validate_rubric_migration, test_run_banner,
  test_run_banner_cli_integration, test_run_banner_stopped.

One inter-test import path updated (test_cli_display_flag.py imports
_make_stub_orchestrator from test_cli_stdout_contract; both moved
together, so the import updates from tests.unit.* → tests.unit.dx.*).

Remaining coupling — deliberate cross-cutting tests staying under
tests/unit/ pending a per-file split (out of scope here):
- test_run_display.py — mixes Protocol tests (engine) with
  LiveRunDisplay tests (dx). Needs a split.
- test_run_display_cli_integration.py — integration test spanning
  engine + dx.
- test_logging_formatter.py — mostly engine, one THEME reference.
- tests/unit/llm/test_preset_overlay_worker_propagation.py — engine
  test that imports a dx helper; the helper should probably move to
  core in a future refactor.

3458 unit+canonical tests still pass — pure relocation, zero
behaviour change. Closes the "engine + dx test-tree decoupling"
follow-up (#456's scope minus the four cross-cutting files noted
above).
