+++ tests/unit/dx/test_adapter_convert_cli.py
