+++ tests/unit/dx/test_cli_assets.py
