+++ tests/unit/dx/test_cli_budget_integration.py
