+++ tests/unit/dx/test_cli_display.py
