+++ tests/unit/dx/test_cli_grouped_help.py
