+++ tests/unit/dx/test_cli_resume.py
