+++ tests/unit/dx/test_cli_status.py
