+++ tests/unit/dx/test_cli_stdout_contract.py
