+++ tests/unit/dx/test_run_banner_stopped.py
